fix: validate baseURL origin and move demo CSV to src/assets
- auth.ts: filter out missing/null/empty Origin header before using
  as baseURL, preventing URL parse failures in edge cases
- Move demo CSV from e2e/fixtures/ to src/assets/ for app bundle,
  decoupling production code from the test suite directory
- e2e tests continue using their own copy at e2e/fixtures/
